requirements.txt
django-ajax-selects
django-admin-interface
Django==3.1.7
django-suit @ https://github.com/darklow/django-suit/tarball/v2